refactor: Вынести datasource и delegate для меню в отдельный класс
